First go at implementing some sort of event pattern. This is a real work in progress
