+++ src/Altinn.App.Core/Interface/IEFormidlingService.cs
@@ -1,0 +1,8 @@
+using Altinn.Platform.Storage.Interface.Models;
+
+namespace Altinn.App.Core.Interface;
+
+public interface IEFormidlingService
+{
+    public Task SendEFormidlingShipment(Instance instance);
+}
