+++ src/Altinn.App.Core/Process/Elements/FeedbackTask.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using Altinn.App.Core.Invokers;
 using Altinn.App.Core.Models;
 using Altinn.App.Services.Interface;
 
@@ -9,32 +10,35 @@ namespace Altinn.App.Core.Process
     /// </summary>
     public class FeedbackTask : TaskBase
     {
-        private readonly IAltinnApp _altinnApp;
+        private readonly IAppEventOrchestrator _orchestrator;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="FeedbackTask"/> class.
         /// </summary>
-        public FeedbackTask(IAltinnApp altinnApp)
+        public FeedbackTask(IAppEventOrchestrator orchestrator)
         {
-            _altinnApp = altinnApp;
+            _orchestrator = orchestrator;
         }
 
         /// <inheritdoc/>
         public override async Task HandleTaskAbandon(ProcessChangeContext processChangeContext)
         {
-            await _altinnApp.OnAbandonProcessTask(processChangeContext.ElementToBeProcessed, processChangeContext.Instance);
+            //await _altinnApp.OnAbandonProcessTask(processChangeContext.ElementToBeProcessed, processChangeContext.Instance);
+            await _orchestrator.OnAbandonProcessTask(processChangeContext.ElementToBeProcessed, processChangeContext.Instance);
         }
 
         /// <inheritdoc/>
         public override async Task HandleTaskComplete(ProcessChangeContext processChangeContext)
         {
-            await _altinnApp.OnEndProcessTask(processChangeContext.ElementToBeProcessed, processChangeContext.Instance);
+            //await _altinnApp.OnEndProcessTask(processChangeContext.ElementToBeProcessed, processChangeContext.Instance);
+            await _orchestrator.OnEndProcessTask(processChangeContext.ElementToBeProcessed, processChangeContext.Instance);
         }
 
         /// <inheritdoc/>
         public override async Task HandleTaskStart(ProcessChangeContext processChangeContext)
         {
-            await _altinnApp.OnStartProcessTask(processChangeContext.ElementToBeProcessed, processChangeContext.Instance, processChangeContext.Prefill);
+            await _orchestrator.OnStartProcessTask(processChangeContext.ElementToBeProcessed, processChangeContext.Instance, processChangeContext.Prefill);
+            //await _altinnApp.OnStartProcessTask(processChangeContext.ElementToBeProcessed, processChangeContext.Instance, processChangeContext.Prefill);
         }
     }
 }
